chore: improved error handling

+++ src/services/user.service.ts
@@ -1,6 +1,7 @@
 import { AppDataSource } from '../config/data-source';
 import { User } from '../entities/user.entity';
 import { UserDto } from '../dtos/user/user.dto';
+import { createHttpError } from '../utils/http-error';
 
 const userRepository = AppDataSource.getRepository(User);
 
@@ -15,16 +16,19 @@ export const findAll = async () => {
 
 export const findOne = async (id: number) => {
   const user = await userRepository.findOneBy({ id });
+  if (!user) throw createHttpError(404, 'User not found', { id });
   return user;
 };
 
 export const update = async (id: number, data: UserDto) => {
-  const user = await findOne(id); 
+  const user = await findOne(id);
+  if (!user) throw createHttpError(404, 'User not found', { id });
   Object.assign(user, data);
   return await userRepository.save(user);
 };
 
 export const remove = async (id: number) => {
   const user = await findOne(id);
+  if (!user) throw createHttpError(404, 'User not found', { id });
   return await userRepository.remove(user);
 };
